Extract `WaitForFile` out of `files.cpp`
`files.cpp` is a large file.

Bug: b/434282027
